Added utils module, containing a simple "freq" function, and a ngram_counter
`freq` simply counts how many times an item appeared in a sequence.
It's roughly equivalent to simply constructing Counter(seq).

ngram_counter counts n-tuples of letters (that may overlap).

E.g. the word "assassin" contains the following 2-tuples:
    as: 2, ss: 2, sa: 1, si: 1, in: 1
When done with 3-tuples:
    ass: 2, ssa: 1, sas, 1 ssi: 1, sin: 1

Characters not in the alphabet will terminate sequences.
E.g. the phrase "to you" contains the following 2-tuples:
    to: 1, yo: 1, ou: 1
And when done with 3-tuples, simply:
    you: 1